+++ audiomate/corpus/io/__init__.py
@@ -26,6 +26,7 @@ from .tatoeba import TatoebaDownloader, TatoebaReader  # noqa: F401
 from .common_voice import CommonVoiceReader  # noqa: F401
 from .mailabs import MailabsDownloader, MailabsReader  # noqa: F401
 from .rouen import RouenDownloader, RouenReader  # noqa: F401
+from .audio_mnist import AudioMNISTDownloader, AudioMNISTReader  # noqa: F401
 
 __downloaders = {}
 for cls in CorpusDownloader.__subclasses__():